update tests for getCatalogFromProgramID
